% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getDataRotNorm2}
\alias{getDataRotNorm2}
\title{Generates data of rotating normal (two-dimensional) distributions.}
\usage{
getDataRotNorm2(angle, n.obs = 1, sd = sqrt(0.2), both = FALSE, time)
}
\arguments{
\item{angle}{\code{[numeric(1)]}\cr
Rotation angle of the means in relation to the y axis (clockwise).}

\item{n.obs}{\code{[numeric(1)]}\cr
Number of observations to draw from the "concept" (distribution).}

\item{sd}{\code{[numeric(1)]}\cr
Standard deviation of the normal distributions.}

\item{both}{\code{[logical(1)]}\cr
If TRUE, observations are equally distributed to both classes.
Default is FALSE.}

\item{time}{\code{[numeric]}\cr
The time stamp to attach to the data.
If missing, it's the "angle" argument.}
}
\value{
\code{[list]}\cr
}
\description{
Generates data of rotating normal (two-dimensional) distributions.
}
\examples{
set.seed(1909)
dat <- getDataRotNorm2(45, 100)
plot(dat$x, col = dat$class)

}

