% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBCD.R
\name{getPredictionModel}
\alias{getPredictionModel}
\title{Get a prediction model from NBCD model.}
\usage{
getPredictionModel(model, pred.time, use.lm = TRUE, n.models = Inf)
}
\arguments{
\item{model}{[NBCD]
NBCD model.}

\item{pred.time}{[numeric(1)]
Time point of prediction.}

\item{use.lm}{[logical(1)]
Use lm models to forecast the movements of mean and sd? (Default TRUE)
If FALSE, use mean and sd of old.model.}

\item{n.models}{[numeric(1)]
Number of past model values for the linear prediction.}
}
\value{
[nb2]
  nb2 model
}
\description{
Get a prediction model from NBCD model.
}

