% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getWaitingTime}
\alias{getWaitingTime}
\title{Wrapper for kMcpp.}
\usage{
getWaitingTime(x, max.len = 10)
}
\arguments{
\item{x}{Probabilities for which to calculate the mean waiting time.}

\item{max.len}{Controls calculation time. If x is longer than max.len, the waiting time will
be approximated: x then just contains the (max.len - 1) smallest values and
the sum of all others. Default is 10.}
}
\value{
[numeric(1)]
Mean waiting time.
}
\description{
Wrapper for kMcpp.
}
\examples{
# Throw a die - how long till every number...
getWaitingTime(rep(1/6, 6))

# Approximation:
\dontrun{
set.seed(1909)
tmp <- probnorm(runif(18))
system.time(print(getWaitingTime(tmp, max.len = 18)))
system.time(print(getWaitingTime(tmp, max.len = 10)))
}

}

