% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getWaitingTime2}
\alias{getWaitingTime2}
\title{Waiting Time Calculation via Integral}
\usage{
getWaitingTime2(p)
}
\arguments{
\item{p}{\code{[numeric]}\cr
Probabilities for which to calculate the mean waiting time.}
}
\value{
\code{[numeric(1)]} Mean waiting time.
}
\description{
The value of this function answers the question:\cr\cr
"How long do I have to wait (on average) until every
value of a discrete distribution (with a finite set of values) appears
\emph{at least} once?"\cr\cr
Uses the solution of Flajolet, Gardy and Thimonier (1992)
}
\examples{
# How long until every side of a fair 40-sided dice appears at least once (on average)?
getWaitingTime2(rep(1/40, 40))


}
\references{
Flajolet, Philippe, Daniele Gardy und Loys Thimonier. "Birthday
Paradox, Coupon Collectors, Caching Algorithms and Self-Organizing Search".
Discrete Applied Mathematics 39(3) (1992), 207--229.
\href{http://dx.doi.org/10.1016/0166-218X(92)90177-C}{DOI:10.1016/0166-218X(92)90177-C}.
}

