% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBCD.R
\name{makeNBCDmodel}
\alias{makeNBCDmodel}
\title{Make or update NBCD model.}
\usage{
makeNBCDmodel(new.obs, model, max.waiting.time, init.obs, verbose = FALSE,
  ..., waiting.time = c("fixed", "auto"), min.waiting.time = 3,
  wait.all.classes = TRUE)
}
\arguments{
\item{new.obs}{[list]
Has elements "x" [data.frame] and "class" [factor]
(and "time", if missing: +1)}

\item{model}{[list]
Model from makeNBCDmodel or empty.}

\item{max.waiting.time}{[numeric(1)]}

\item{init.obs}{[numeric(1)]}

\item{verbose}{[logical(1)]}

\item{...}{[any]
Arguments passed to update.nb2(...).}

\item{waiting.time}{[character(1)]
Default is "fixed". If "auto", then calculate waiting time between two
models (for factorial variables) with getWaitingTime(). If "fixed", use
max.waiting.time.}

\item{min.waiting.time}{[numeric(1)]
Minimal waiting time, if waiting.time = "auto".}

\item{wait.all.classes}{[logical(1)]
If waiting.time = "auto", should algo wait for all classes to be appeared?}
}
\value{
[NBCD]
NBCD model
}
\description{
Make or update NBCD model.
}

