% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveBayes.R
\name{nb2}
\alias{nb2}
\title{nb2: Extended Naive Bayes - Updateable Version}
\usage{
nb2(x, y, laplace = 0, discretize, discParams, ...)
}
\arguments{
\item{x}{[matrix | data.frame]
A numeric matrix, or a data frame of categorical and/or numeric variables.}

\item{y}{[numeric | factor | character]
Class vector.}

\item{laplace}{[numeric(1)]
positive double controlling Laplace smoothing.
The default (0) disables Laplace smoothing.}

\item{discretize}{[character(1)]
If missing: No discretization will be performed and discParams argument is
  ignored.
If "fixed": number of intervals or there boundaries have to be given in the
  discParam argument}

\item{discParams}{[numeric | list]
If discretize = "fixed": Has to be a named list giving the boundaries for
  the variables to be discretized.}

\item{...}{Currently not used.}
}
\value{
[list] of type "naiveBayes"
additional element: "yc" contains Youngs-Cramer variance parameters
}
\description{
nb2: Extended Naive Bayes - Updateable Version
}
\examples{
mod <- nb2(iris[, 1:4], iris[, 5])
mod

discParam <- list(Sepal.L = c(5, 6, 7), Sepal.W = c(2.5, 3.5))
mod2 <- nb2(iris[, 1:4], iris[, 5], discretize = "fixed", discParams = discParam)
mod2

}

