% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pastehead}
\alias{pastehead}
\title{Paste head of object to a string with a maximal length.}
\usage{
pastehead(x, len = 3, coll = " / ", end = "...")
}
\arguments{
\item{x}{[numeric | character]
Object to paste.}

\item{len}{[numeric]
Maximal length.}

\item{coll}{[character]
Separation string.}

\item{end}{[character]
End string.}
}
\value{
[character(1)]
}
\description{
Paste head of object to a string with a maximal length.
}

