% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBCD.R
\name{plot.NBCD}
\alias{plot.NBCD}
\title{Plot predicted class values of NBCD model.}
\usage{
\method{plot}{NBCD}(x, time, use.lm = FALSE, ..., n.models = Inf,
  .verb = TRUE)
}
\arguments{
\item{x}{[NBCD]
NBCD model.}

\item{time}{[numeric(1)]
Time for which to predict (passed to predict.NBCD(...)).}

\item{use.lm}{[logical(1)]
Use lm models to forecast the movements of mean and sd? (Default TRUE)
If FALSE, use mean and sd of old.model.
(passed to predict.NBCD(...))}

\item{...}{[any]
Arguments passed to plot.nb2(...).}

\item{n.models}{[numeric(1)]
Number of past model values for the linear prediction.}

\item{.verb}{[logical(1)]
Show messages? Default TRUE.}
}
\description{
Plot predicted class values of NBCD model.
}

