% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBCD.R
\name{predict.NBCD}
\alias{predict.NBCD}
\title{Predict class values for new data with NBCD model.}
\usage{
\method{predict}{NBCD}(object, newdata, time, use.lm, n.models = Inf, ...)
}
\arguments{
\item{object}{[NBCD]
NBCD model.}

\item{newdata}{[data.frame]
Containing variables.}

\item{time}{[numeric(1)]
Time for which to predict.}

\item{use.lm}{[logical(1)]
Use lm models to forecast the movements of mean and sd? (Default TRUE)
If FALSE, use mean and sd of old.model.}

\item{n.models}{[numeric(1)]
Number of past model values for the linear prediction.}

\item{...}{Arguments passed to predict.nb2(...).}
}
\value{
Results from predict.nb2(...).
}
\description{
Predict class values for new data with NBCD model.
}

