% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveBayes.R
\name{print.nb2}
\alias{print.nb2}
\title{Printing nb2 objects.}
\usage{
\method{print}{nb2}(x, ..., digits = getOption("digits"), print.apl = FALSE)
}
\arguments{
\item{x}{[nb2]
nb2 object to print.}

\item{...}{Currently not used.}

\item{digits}{[numeric]}

\item{print.apl}{[logical]
Print apriori.list? Else: apriori.}
}
\description{
Modified copy of e1071:::print.naiveBayes.
}

