% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBCD.R
\name{print.NBCD}
\alias{print.NBCD}
\title{Print information about NBCD model.}
\usage{
\method{print}{NBCD}(x, size = c("small", "big"), ..., use.lm = FALSE)
}
\arguments{
\item{x}{[NBCD]
NBCD model.}

\item{size}{[character]
small or big}

\item{...}{[any]
Currently ignored.}

\item{use.lm}{[logical]
use lm for prediction?}
}
\description{
Print information about NBCD model.
}

