% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{probnorm}
\alias{probnorm}
\title{"Normalize" objects.}
\usage{
probnorm(x)
}
\arguments{
\item{x}{[numeric]
Object}
}
\value{
Object divided by sum of its values.
}
\description{
"Normalize" objects.
}
\examples{
probnorm(1:5)

}

