% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBCD.R
\name{resetModel}
\alias{resetModel}
\title{Reset Model when waiting.time reached}
\usage{
resetModel(model, var.name)
}
\arguments{
\item{model}{\code{[NBCD]}\cr
"model$current"}

\item{var.name}{\code{[character]}\cr}
}
\value{
\code{[NBCD]}\cr
  NBCD model with resetted "current" part.
}
\description{
Reset Model when waiting.time reached
}

