% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rnhyper}
\alias{rnhyper}
\title{Create hyperplane data with independant normal distributions}
\usage{
rnhyper(n, mean = 0, sd = 1, d)
}
\arguments{
\item{n}{[numeric(1)]}

\item{mean}{[numeric(1) | numeric(d)]}

\item{sd}{[numeric(1) | numeric(d) | matrix(d, d)]
If matrix, it's a VCV with variances on the diagonal.
Else, interpreted as standard deviations.}

\item{d}{[numeric(1)]
Dim.}
}
\description{
Create hyperplane data with independant normal distributions
}

