% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{updateVarYC}
\alias{updateVarYC}
\title{Variance calculation (updatable) by Youngs and Cramer}
\usage{
updateVarYC(x, object = NULL)
}
\arguments{
\item{x}{[numeric]
vector (finite, no missing values)}

\item{object}{[list | NULL]
list from updateVarYC or NULL}
}
\value{
[list(1)]
  $var: variance of x via Youngs-Cramer-algorithm
}
\description{
Variance calculation (updatable) by Youngs and Cramer
}
\section{Alternatives}{

\itemize{
  \item Cotton, E.W. (1975), Remark on "Stably Updating Mean and Standard
   Deviation of Data," Communications of the Association for Computing
   Machinery, 18, 458.
  \item Hanson, R.J. (1975), "Stably Updating Mean and Standard Deviation of
   Data," Communications of the Association for Computing Machinery, 18,
   57-58.
  \item Welford, B.P. (1962), "Note on a Method for Calculating Corrected Sums
   of Squares and Products," Technometrics, 4, 419-420.
  \item West, D.H.D. (1979), "Updating Mean and Variance Estimates: An Improved
   Method," Communications of the Association for Computing Machinery,
   22, 532-535.
}
}
\examples{
\dontrun{
  yc.fun = function(x) {
    yc = NULL
    for (i in seq_along(x)) {
      yc = NBCD:::updateVarYC(x[i], yc)
    }
    return(yc$var)
  }
  yc.fun(1:10)
}

}
\references{
Youngs, Edward A., and Elliot M. Cramer. "Some results relevant to choice of
sum and sum-of-product algorithms." Technometrics 13.3 (1971): 657--665,
\href{http://dx.doi.org/10.2307/1267176}{DOI:10.2307/1267176}.
See also:
\url{http://scholar.google.de/scholar?cluster=10728821699746172437&hl=de}
}

