% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBCD.R
\name{getPredictionModel}
\alias{getPredictionModel}
\title{Get a prediction model from NBCD model.}
\usage{
getPredictionModel(model, pred.time, use.lm = TRUE, n.models = Inf)
}
\arguments{
\item{model}{\code{[NBCD]}\cr
NBCD model.}

\item{pred.time}{\code{[numeric(1)]}\cr
Time point of prediction.}

\item{use.lm}{\code{[logical(1)]}\cr
Use lm models to forecast the movements of mean and sd? (Default TRUE)
If FALSE, use mean and sd of old.model.}

\item{n.models}{\code{[numeric(1)]}\cr
Number of past model values for the linear prediction.}
}
\value{
\code{[nb2]}\cr
  nb2 model
}
\description{
Get a prediction model from NBCD model.
}

