% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBCD.R
\name{makeNBCDmodel}
\alias{makeNBCDmodel}
\title{Make or update NBCD model.}
\usage{
makeNBCDmodel(new.obs, model, max.waiting.time, init.obs, verbose = FALSE,
  ..., waiting.time = c("fixed", "auto"), min.waiting.time = 3,
  wait.all.classes = TRUE)
}
\arguments{
\item{new.obs}{\code{[list]}\cr
Has elements "x" \code{[data.frame]} and "class" \code{[factor]}
(and "time", if missing: +1)}

\item{model}{\code{[list]}\cr
Model from makeNBCDmodel or empty.}

\item{max.waiting.time}{\code{[numeric(1)]}\cr}

\item{init.obs}{\code{[numeric(1)]}\cr}

\item{verbose}{\code{[logical(1)]}\cr}

\item{...}{\code{[any]}\cr
Arguments passed to update.nb2(...).}

\item{waiting.time}{\code{[character(1)]}\cr
Default is "fixed". If "auto", then calculate waiting time between two
models (for factorial variables) with getWaitingTime2(). If "fixed", use
max.waiting.time.}

\item{min.waiting.time}{\code{[numeric(1)]}\cr
Minimal waiting time, if waiting.time = "auto".}

\item{wait.all.classes}{\code{[logical(1)]}\cr
If waiting.time = "auto", should algo wait for all classes to be appeared?}
}
\value{
\code{[NBCD]}\cr
NBCD model
}
\description{
Make or update NBCD model.
}

