% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pastehead}
\alias{pastehead}
\title{Paste head of object to a string with a maximal length.}
\usage{
pastehead(x, len = 3, coll = " / ", end = "...")
}
\arguments{
\item{x}{\code{[numeric | character]}\cr
Object to paste.}

\item{len}{\code{[numeric]}\cr
Maximal length.}

\item{coll}{\code{[character]}\cr
Separation string.}

\item{end}{\code{[character]}\cr
End string.}
}
\value{
\code{[character(1)]}\cr
}
\description{
Paste head of object to a string with a maximal length.
}

