% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveBayes.R
\name{predict.nb2}
\alias{predict.nb2}
\title{Prediction for Extended Naive Bayes (nb2)}
\usage{
\method{predict}{nb2}(object, newdata, type = c("class", "raw"),
  threshold = 0.001, eps = 0, ...)
}
\arguments{
\item{object}{An object of class "nb2".}

\item{newdata}{A dataframe with new predictors (with possibly fewer columns than
the training data). Note that the column names of newdata are matched
against the training data ones.}

\item{type}{If "raw", the conditional a-posterior probabilities for each class are
returned, and the class with maximal probability else.}

\item{threshold}{Value replacing cells with probabilities within eps range.}

\item{eps}{double for specifying an epsilon-range to apply laplace smoothing
(to replace zero or close-zero probabilities by theshold.)}

\item{...}{Currently ignored.}
}
\description{
Prediction for Extended Naive Bayes (nb2)
}
\examples{
mod <- nb2(iris[, 1:4], iris[, 5])
predict(mod, iris)

discParam <- list(Sepal.L = c(5, 6, 7), Sepal.W = c(2.5, 3.5))
mod2 <- nb2(iris[, 1:4], iris[, 5], discretize = "fixed", discParam = discParam)
predict(mod2, iris)

c(mean(predict(mod, iris) == iris$Species),
  mean(predict(mod2, iris) == iris$Species))

}

