% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBCD.R
\name{predict.NBCD}
\alias{predict.NBCD}
\title{Predict class values for new data with NBCD model.}
\usage{
\method{predict}{NBCD}(object, newdata, time, use.lm, n.models = Inf, ...)
}
\arguments{
\item{object}{\code{[NBCD]}\cr
NBCD model.}

\item{newdata}{\code{[data.frame]}\cr
Containing variables.}

\item{time}{\code{[numeric(1)]}\cr
Time for which to predict.}

\item{use.lm}{\code{[logical(1)]}\cr
Use lm models to forecast the movements of mean and sd? (Default TRUE)
If FALSE, use mean and sd of old.model.}

\item{n.models}{\code{[numeric(1)]}\cr
Number of past model values for the linear prediction.}

\item{...}{Arguments passed to predict.nb2(...).}
}
\value{
Results from predict.nb2(...).
}
\description{
Predict class values for new data with NBCD model.
}

