% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveBayes.R
\name{print.nb2}
\alias{print.nb2}
\title{Printing nb2 objects.}
\usage{
\method{print}{nb2}(x, ..., digits = getOption("digits"), print.apl = FALSE)
}
\arguments{
\item{x}{\code{[nb2]}\cr
nb2 object to print.}

\item{...}{Currently not used.}

\item{digits}{\code{[numeric]}\cr}

\item{print.apl}{\code{[logical]}\cr
Print apriori.list? Else: apriori.}
}
\description{
Modified copy of \link[e1071]{print.naiveBayes} from \code{e1071}.
}
\author{
\itemize{
  \item David Meyer \email{David.Meyer@R-project.org}: original \link[e1071]{print.naiveBayes}
  \item Philipp Aschersleben \email{aschersleben@statistik.tu-dortmund.de}:
  modifications and extensions for the NBCD package
}
}
\seealso{
\code{\link{nb2}}, \code{\link{update.nb2}}, \code{\link{plot.nb2}},
\code{\link{predict.nb2}}
}

