% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBCD.R
\name{print.NBCD}
\alias{print.NBCD}
\title{Print information about NBCD model.}
\usage{
\method{print}{NBCD}(x, size = c("small", "big"), ..., use.lm = FALSE, time,
  len = 3)
}
\arguments{
\item{x}{\code{[NBCD]}\cr
NBCD model.}

\item{size}{\code{[character]}\cr
small or big}

\item{...}{\code{[any]}\cr
Currently ignored.}

\item{use.lm}{\code{[logical]}\cr
use lm for prediction?}

\item{len}{\code{[numeric(1)]}\cr
max. number of printed names/numbers/...}
}
\description{
Print information about NBCD model.
}

