% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rnhyper}
\alias{rnhyper}
\title{Create hyperplane data with independant normal distributions}
\usage{
rnhyper(n, mean = 0, sd = 1, d)
}
\arguments{
\item{n}{\code{[numeric(1)]}\cr}

\item{mean}{\code{[numeric(1) | numeric(d)]}\cr}

\item{sd}{\code{[numeric(1) | numeric(d) | matrix(d, d)]}\cr
If matrix, it's a VCV with variances on the diagonal.
Else, interpreted as standard deviations.}

\item{d}{\code{[numeric(1)]}\cr
Dim.}
}
\description{
Create hyperplane data with independant normal distributions
}

